[proc macros]: remove old code and tests. (#431)
* fix: update example to new proc macros

* remove unused code

* remove client builder
